Separate point, tripoint into new point.h
It never made sense that point and tripoint were in enums.h.  Give them
their own header.
