+++ src/advanced_inv.h
@@ -13,6 +13,7 @@
 
 #include "cursesdef.h"
 #include "enums.h"
+#include "point.h"
 #include "units.h"
 
 class uilist;
