+++ src/coordinate_conversions.h
@@ -3,6 +3,7 @@
 #define COORDINATE_CONVERSIONS_H
 
 #include "enums.h"
+#include "point.h"
 
 /**
  * Coordinate systems used here are:
