+++ src/creature_tracker.h
@@ -8,6 +8,7 @@
 #include <vector>
 
 #include "enums.h"
+#include "point.h"
 
 class monster;
 class JsonIn;
