+++ src/editmap.h
@@ -10,6 +10,7 @@
 #include "color.h"
 #include "cursesdef.h"
 #include "enums.h"
+#include "point.h"
 #include "type_id.h"
 
 struct real_coords;
