+++ src/magic_teleporter_list.h
@@ -8,6 +8,7 @@
 
 #include "enums.h"
 #include "optional.h"
+#include "point.h"
 
 class avatar;
 class JsonIn;
