+++ src/map_selector.h
@@ -5,6 +5,7 @@
 #include <vector>
 
 #include "enums.h"
+#include "point.h"
 #include "visitable.h"
 
 class map_cursor : public tripoint, public visitable<map_cursor>
