+++ src/mapbuffer.h
@@ -8,6 +8,7 @@
 #include <string>
 
 #include "enums.h"
+#include "point.h"
 
 class submap;
 class JsonIn;
