+++ src/overmap_noise.h
@@ -4,6 +4,7 @@
 
 #include "enums.h"
 #include "game_constants.h"
+#include "point.h"
 
 namespace om_noise
 {
