+++ src/overmap_ui.h
@@ -3,6 +3,7 @@
 #define OVERMAP_UI_H
 
 #include "enums.h"
+#include "point.h"
 
 namespace catacurses
 {
