+++ tests/map_helpers.h
@@ -4,7 +4,7 @@
 
 #include <string>
 
-#include "enums.h"
+#include "point.h"
 
 class monster;
